requirements.txt
pyttsx3
SpeechRecognition
google-generativeai
pyjokes
wikipedia-api
